Ran the code through flycheck & checkdoc
Made sure that the doc strings conform to the correct format and that
the static code analysis passes.

Then removed sort-words-list-to-string and added sort-words-sorted as it
made more sense to return the sorted word list as a string from
sort-words-sorted function.

I'm using sort-words-in-region & sort-words-sorted because I don't want
to put all this functionality all into sort-words function, because I
think this way is easier to read.